I ended up writing a blog post about the tools I used for profiling the change I proposed to Occ::get. Fun story, @aatch [pointed out on Reddit](https://www.reddit.com/r/rust/comments/4udxmr/rust_performance_a_story_featuring_perf_and/d5plly1) that the slowdown in some other versions I had tried was probably due to the overhead of dealing with alignment in a vectorized version of the function, and noted that 3 may be an odd sample rate to test with the FMIndex. As an example, bowtie2's FM index appears to default to a sampling rate of 32, and I've seen a few cases of using higher rates as well.

The version of Occ::get proposed by Aatch is at approximate parity with the filter/count version for performance at a sampling rate of 32 (just a tiny bit slower), but it quickly gets much, much faster as the sampling rate increases. For example, with a sampling rate of 128 (in use in one of my applications):

filter/count, k = 128

```
test search_index_seeds ... bench:     179,442 ns/iter (+/- 1,947)
```

vectorized, k = 128

```
test search_index_seeds ... bench:     152,368 ns/iter (+/- 433)
```

filter/count, k = 64

```
test search_index_seeds ... bench:     105,141 ns/iter (+/- 1,665)
```

vectorized, k = 64

```
test search_index_seeds ... bench:      99,044 ns/iter (+/- 252)
```

filter/count, k = 32

```
test search_index_seeds ... bench:      74,427 ns/iter (+/- 5,166)
```

vectorized, k = 32

```
test search_index_seeds ... bench:      75,333 ns/iter (+/- 303)
```

Seems like an overall win to me. At a `k` of 32, the averages are just slightly higher than the filter/count version, but well within the error margin reported by libtest.

The one downside here is that at smaller sampling rates, a non-vectorized version wins out:

filter/count, k = 8

```
test search_index_seeds ... bench:      40,598 ns/iter (+/- 1,170)
```

vectorized, k = 8

```
test search_index_seeds ... bench:      53,842 ns/iter (+/- 224)
```

Thoughts here? Without integer generics (at all) and specialization (in stable), I'm not sure it's possible to cleanly maximize performance of small sampling rates and large ones. I'm biased here because I'm working with very large indices that need high sampling rates, but I'm not sure what other values of `k` are "in the wild."