@Aatch 2016-07-27
One change I'm investigating is making `k` a `u32` instead of a `usize`. On a 64-bit platform, with the default settings, the way the vectorized code works means that it is limited to processing 4 elements at a time, but limit `k` to a `u32` allows it to process 8 elements at a time. My testing shows that for `k == 32`, this increases speed for the loop over the non-vectorized version by a reasonable amount.

For `k == 32`:

```
filter-count:
   test search_index_seeds ... bench:      33,397 ns/iter (+/- 1,498)

loop:
   test search_index_seeds ... bench:      37,146 ns/iter (+/- 1,266)

loop, k: u32:
   test search_index_seeds ... bench:      31,221 ns/iter (+/- 1,982)
```

For `k == 64`

```
filter-count:
   test search_index_seeds ... bench:      49,810 ns/iter (+/- 9,850)

loop, k: u32
   test search_index_seeds ... bench:      33,503 ns/iter (+/- 1,824)
```

And for smaller `k`, it is still is slower, I don't think it's a real problem, since 32/64 are apparently the more common sizes.

```
filter-count:
   test search_index_seeds ... bench:      18,364 ns/iter (+/- 1,374)

loop, k: u32:
   test search_index_seeds ... bench:      27,241 ns/iter (+/- 1,781)
```

At this point, you do need a more in-depth benchmark to see the impact, as the whole point of this structure is to control the space/time tradeoff for counting occurrences. Cache effects and even paging are going to start being relevant.