@Aatch 2016-07-27
@dikaiosune it only limits the size of the "chunks" the input is split into, not the size of the data itself. Given that you're currently splitting into 128-byte chunks, the fact that using `u32` limits you to 4GB chunks doesn't seem too ridiculous. The reason for limiting `k` here is actually so you can use a `u32` for the `count` variable instead of a `usize`, meaning that instead of the vectorized loop using two `<2 x u64>` vectors, it can use two `<4 x u32>` vectors and process 8 elements at once instead.