@dikaiosune 2016-07-27
@Aatch If we use a u32 for `count`, then counting >4B matching characters from the BWT will overflow, which is likely to happen when counting up to the final rows in very large BWT's. (I think)