@dikaiosune 2016-07-27
Note for later: for texts which are small enough, the same optimization could apply nicely to the SuffixArray impl as well. SeqAn lets you [pick the type of the suffix array items](https://github.com/seqan/seqan/blob/master/manual/attic/Tutorial/Indices/FMIndex.rst#reducing-the-memory-consumption) so you can take advantage of 32-bit speed on <4GB texts.