@Aatch 2016-07-27
@dikaiosune it won't. The maximum value of `count` (the "manual count") is `k - 1`, since you only process `k - 1` elements. You then add that to the stored count for the beginning of the block you're in. That stored count should continue to be a `usize`, you just cast the "manual count" to a usize before adding it to the stored count.