@dikaiosune 2016-07-27
So, to clarify, you're suggesting something like:

```rust
    pub fn get(&self, bwt: &BWTSlice, r: usize, a: u8) -> usize {
        // self.k is our sampling rate, so find our last sampled checkpoint
        let i = r / self.k;
        let checkpoint = self.occ[i][a as usize];

        // find the portion of the BWT past the checkpoint which we need to count
        let start = (i * self.k) + 1;
        let end = r + 1;

        // count all the matching bytes b/t the closest checkpoint and our desired lookup
        let mut count: u32 = 0;
        for &x in &bwt[start..end] {
            if x == a {
                count += 1;
            }
        }

        // return the sampled checkpoint for this character + the manual count we just did
        checkpoint + (count as usize)
    }
```

Is that accurate? Why does `k` also need to be a u32 here? I think that <4B sampling rates are completely reasonable, just curious.