@Aatch 2016-07-27
@dikaiosune `k` only needs to be a `u32` to ensure that `count` doesn't overflow.