@johanneskoester 2016-08-15
Ok, I'm fine with u32 and a custom loop then. Thanks a lot for evaluating this!!